Add configuration values to spec helper so tests can run out of the box
Normally config values are loaded from config/config.yml, which would
require this to be checked in. But, we manage that file with chef.
This change allows the spec tests to be automated and those new to this
project to not get hung up on not having the config file for tests.