Added backup handling of updating existing backup.
So, you used the backup utility to pull everything down, but then a new nightly was released. With this patch, it will not re-download everything, but only download files where the bottom level directory is missing. 

Therefore, works well until you fail downloading a file. In that case, you should just nuke the entire version directory to be on the safe side and re-run.