uniq packages by relpath when finding expected checksums
we reuse packages on different platform versions to reduce the number of
packages we need to build. When getting the list of expected checksums,
we need to uniq the expected metadata objects by package relpath or else
we overcount the number of expected packages, and the verifier reports a
negative number of unverifiable packages.