Fix crash caused by malformed version numbers
Stack trace of the crash:
------------------------

```
NoMethodError - undefined method `release?' for #<Array:0x007ff24a604d68>:
  /Users/ddeleo/oc/opscode-omnitruck/lib/opscode/version.rb:296:in `block in find_target_version'
  /Users/ddeleo/oc/opscode-omnitruck/lib/opscode/version.rb:276:in `select'
```

Data that causes crash:
----------------------

```json
{
  "el": {
    "5": {
      "x86_64": {
        "": "/el/5/x86_64/chef-server-11.0.8+20130510085727.git.10.115aaf4-1.el5.x86_64.rpm",
```

Bug
---

The existing code that checks for malformed versions iterates over a list of
version numbering types takes the return value from Array#each, and uses
`break` within the loop to return a single value. When a version string
is not a valid version of any of the types, the variable was set to the
return of `Array#each`, which is the array. The test for
non-matching-ness expected a non-match to return nil.