Added quotes to wget and curl args to address CHEF-4080 (see below).
P.S. Could not repro the above issue in bash or dash. If you run $wget some-url-with-& it will break if not quoted; however, if you set the variable url=some-url-with-& then wget $url, it does not need the quote escaping as far as I can tell.

Maybe it was an early bug in dash that was patched later?