Get filename and type from metadata, drop `-s` switch
Use the actual package filename and get the filetype from its suffix.
This ensures that we can safely cache the packages as the name includes
all platform and version information. And previously the `version`
variable was even empty by default.

Also drop the obsolete `use_shell` option as it won't make sense to try
to feed other then .sh files to the shell.