skip broken versions
if we push a bad metadata version up to s3 we will not be able to parse
it and the exception that is thrown will break all downloads and will
cause an outage to omnitruck (if it occurred in production).  since this
is a terrible outcome, we just skip versions that don't parse.