Workaround ARM package issues on Fedora ARM.
While the rpm itself is genated on Fedora, the resulting RPM cannot
be installed as rpm complains about the arch. Interestingly enough,
attempting to change the arch to what RPM expects when attempting to
build the omnibus RPM results in rpmbuild complaining that armv7hl,
the expected arch, is an invalid architecture.

While this is not the ideal solution, I do not beleive ohai is
incorrect in the result it obtains for the architecture. As such,
this issue seems to be isolated to just installing the omnibus
packages.

Closes #120