Allow specifying projects for s3_poller.
In all use cases except one we want to poll every possible project. The
one exception being when a new platform or architecture is being
developed. The reason being omnibus-chef builds require either chef, or
angry chef to build out the environment, and releasing them is easier
if it can be done piecemeal rather than as a monolithic release.