Updates cache key to use X-Forwarded-Proto
The original setup of this caching assumed that the scheme would handle
the https/http difference in the cache. Because we are terminating ssl
on the ELB and not on the box, we need to add the X-Forwarded-Proto to
the cache key to ensure we return the correct protocol in our 302.