Cookbook updates for habitat
This commit changes the build cookbook and the omnitruck deployment
cookbook to support building the habitat package and deploying it on
the omnitruck instance. The build-cookbook is simplified in a few ways.

First, we change to installing the aws-sdk gem via the cookbook
metadata. This gem does not have native extensions, and we can remove
that part of the default recipe to make it responsible for less.

Next, we'll use the habitat-build cookbook for publishing. This allows
us to remove the awscli from the default recipe, and remove all the
AWS S3 publishing related code from the publish recipe. We also
convert the inline aws config for chef-provisioning to use a template
instead.

Chef no longer requires templates/default or files/default, so we can
remove those to make it easier to get to the files we manage with Chef.

Finally, we rewrite the default recipe using the resources from the
habitat cookbook to manage the application. We no longer need the
included recipes. They are an unnecessary optimization for this
application. Note taht the two services will run on the same machine -
omnitruck and its nginx proxy. This means we need to run the habitat
supervisor on different ports, hence the unit_content of that
service's resource. The habitat project is working on moving the
functionality of the habitat director to the supervisor, so we're
going to plan ahead by using the supervisor for these services, rather
than managing a director configuration.

Signed-off-by: Joshua Timberman <joshua@chef.io>