Remove hab-director config
The hab-director will no longer be a standalone component upstream in
Habitat, and we won't use it here. Instead we'll use the `listen`
options to select ports for the services running on the same system
(see default recipe), and make use of the "director" features when
they're in the habitat supervisor in the future.

Signed-off-by: Joshua Timberman <joshua@chef.io>