ensure application files are readable by non-root users
Sometimes, gems get released with 0640 permissions, which happened
with the statsd gem v 1.2.1 (.gitignore). Because the application runs
as `hab`, it doesn't have permission to copy this file to the static
directory in the init hook. This change finds all files that don't
have "other read" and changes the permission to allow that.

Signed-off-by: Joshua Timberman <joshua@chef.io>