Removes guard
The assumption the removed call makes is that we would change a habitat
file each time we push through a change. Since the version is generated
dynamically, there really should be no reason to modify those files. In
the case of the application going through the pipeline, we should always
be building a new hart since it is the entire purpose of the pipeline.

Signed-off-by: Christopher Webber <cwebber@chef.io>