pin bundler and ruby versions, create the cache directory
Omnitruck needs ruby 2.3.1, newer versions have a problem with the
json library that we haven't ironed out. We also need to pin the
bundler version because it will pull in a newer ruby if we don't, and
then we get dependency conflicts.

We also need to create the cache directory with the version in the
prepare step, as the variable that the plan build script used has
changed in habitat 0.21.0.

Signed-off-by: Joshua Timberman <joshua@chef.io>