Use TCP sockets for IPC.
This uses TCP sockets rathet than local unix sockets to facilitate
communiucation between docker containers. The one part which does not
work in this model is the `omnitruck-poller`. This is because the
omnitruck application looks for the files which the poller creates
within it's own context. There are two workarounds to this problem:
 1. Create a shared volume for both containers to read/write the poller
artifacts.
 2. Execute the poller application as a cron job from the host from
    within the container. Eg: `docker exec -d container_name hab pkg
exec chef-es/omnitruck bundle exec poller -e production -s 300`

Both of these suggested workarounds are untested, so your milage may
vary.