Send restart to supervisor on upgrade.
We need to send a restart to the supervisor rather than use Habitat
update strategy since we do not currently leverage Habitat channels.
This is needed so we can control updates to different environments
since the hab supervisor does not restart services automatically when a
new version is installed without an update strategy defined.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>