Stop pinning Ruby version
The current build of Omnitruck is failing on startup with the following
error:

```
bundler: command not found: rackup
Install missing gem executables with `bundle install`
```

Omnitruck's Habitat build takes advantage of the Ruby Scaffolding
which defines build patterns for Ruby-based Habitat plans:
https://github.com/habitat-sh/core-plans/blob/master/scaffolding-ruby/doc/reference.md

It appears Ruby Scaffolding makes some assumptions on which version of
Ruby is being used when the `bundle install` command is executed at
build time. We previously pinned to `core/ruby24` but the `bundle
install` was targeting the `core/ruby` install. This meant the bundled
gems where not available at startup time. The easiest fix for us in this
situation is to remove our hard pinning on a specific Ruby version and
to just ride the lastest Ruby as delivered via `core/ruby`. This is also
better from a security perspective as we will always be using the latest
Ruby each time we build.

Signed-off-by: Seth Chisamore <schisamo@chef.io>